Add support of holonomic robot by adding missing Y component in odometry (#71)
* Add support of holonomic robot by adding missing Y component in odometry

The odometry was only computed taking in account linear velocity on the x-axis and angular
velocity around the z-axis. The code is modified to include linear velocity on the y-axis for
support of holonomic robots. Odometry arc and radius are calculated the same way as before but
using the combined linear velocity instead. The end position is then rotated by the angle between
the combined velocity vector and the x-axis

* Change variable name

* Rename combined velocity angle variable