Add case for rotating in place in odometry
Odometry calculations were not taking in account the case of the robot rotating in place. This
would cause a division by zero and make the plugin return invalid data. This commit handles this
case by not using the arc method if the distance traveled by the robot is close to zero.