Fix negative Y-axis motion bug
Only the displacement along X and the combined displacement were used to
calculate the angle between the X-axis and the robot drive direction.
This resulted in a positive angle whatever if the displacement along Y
was positive or negative. This commit addresses the issue by using both
displacements along X and Y to calculate the angle.