Add case for rotating in place in odometry (#75)
* Add case for rotating in place in odometry

Odometry calculations were not taking in account the case of the robot rotating in place. This
would cause a division by zero and make the plugin return invalid data. This commit handles this
case by not using the arc method if the distance traveled by the robot is close to zero.

* Fix negative Y-axis motion bug

Only the displacement along X and the combined displacement were used to
calculate the angle between the X-axis and the robot drive direction.
This resulted in a positive angle whatever if the displacement along Y
was positive or negative. This commit addresses the issue by using both
displacements along X and Y to calculate the angle.